# hub-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EmailDJ Hub API",
    "version": "0.1.0"
  },
  "paths": {
    "/generate/quick": {
      "post": {
        "tags": [
          "generate"
        ],
        "summary": "Start Quick Generate",
        "operationId": "start_quick_generate_generate_quick_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuickGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuickGenerateAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/generate/stream/{request_id}": {
      "get": {
        "tags": [
          "generate"
        ],
        "summary": "Stream Quick Generate",
        "operationId": "stream_quick_generate_generate_stream__request_id__get",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExtractionMetadataIn": {
        "properties": {
          "selectorConfidences": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Selectorconfidences"
          },
          "extractedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extractedat"
          },
          "salesforceUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salesforceurl"
          }
        },
        "type": "object",
        "title": "ExtractionMetadataIn"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ProspectPayload": {
        "properties": {
          "accountId": {
            "type": "string",
            "title": "Accountid"
          },
          "accountName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accountname"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry"
          },
          "employeeCount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employeecount"
          },
          "openOpportunities": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Openopportunities"
          },
          "lastActivityDate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastactivitydate"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          },
          "activityTimeline": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Activitytimeline"
          },
          "extractionMetadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExtractionMetadataIn"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "accountId"
        ],
        "title": "ProspectPayload"
      },
      "QuickGenerateAccepted": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "stream_url": {
            "type": "string",
            "title": "Stream Url"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "stream_url"
        ],
        "title": "QuickGenerateAccepted"
      },
      "QuickGenerateRequest": {
        "properties": {
          "payload": {
            "$ref": "#/components/schemas/ProspectPayload"
          },
          "slider_value": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Slider Value",
            "default": 5
          }
        },
        "type": "object",
        "required": [
          "payload"
        ],
        "title": "QuickGenerateRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}